260406 finish

+++ src/main/java/com/merge/final_project/campaign/campaigns/repository/CampaignRepository.java
@@ -3,5 +3,5 @@ package com.merge.final_project.campaign.campaigns.repository;
 import com.merge.final_project.campaign.campaigns.entity.Campaign;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CampaignRepository extends JpaRepository<Campaign, Integer> {
+public interface CampaignRepository extends JpaRepository<Campaign, Long> {
 }
