+++ src/main/java/com/merge/final_project/campaign/useplan/entity/UsePlan.java
@@ -17,14 +17,14 @@ public class UsePlan {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "use_plan_no")
-    private Integer usePlanNo;
+    private Long usePlanNo;
 
     @Column(name = "plan_content", nullable = false)
     private String planContent;
 
     @Column(name = "plan_amount", nullable = false)
-    private Integer planAmount;
+    private Long planAmount;
 
     @Column(name = "campaign_no", nullable = false)
-    private Integer campaignNo;
+    private Long campaignNo;
 }
